requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
requests==2.31.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.107.0
aiohttp==3.8.6
pydantic==2.4.2 